@@ -1,23 +1,15 @@
 {
-  "name": "team-mates-tests",
+  "name": "tests",
   "private": true,
-  "version": "0.0.1",
-  "description": "",
-  "main": "index.js",
   "scripts": {
-    "test": "TRYORAMA_LOG_LEVEL=info WASM_LOG=warn RUST_LOG=warn RUST_BACKTRACE=1 TRYORAMA_HOLOCHAIN_PATH=\"holochain\" ts-node src/index.ts"
+    "test": "vitest run"
   },
-  "author": "",
-  "license": "ISC",
   "dependencies": {
-    "@holochain-open-dev/core-types": "0.1.1",
-    "@holochain/tryorama": "0.4.6",
-    "@types/lodash": "^4.14.168",
-    "@types/node": "^14.14.31",
-    "concurrently": "^6.0.0",
-    "lodash": "^4.17.21",
-    "tape": "^5.2.0",
-    "ts-node": "^9.1.1",
-    "typescript": "4.3.5"
-  }
+    "@msgpack/msgpack": "^2.7.0",
+    "@holochain/client": "^0.12.2",
+    "@holochain/tryorama": "^0.11.0",
+    "typescript": "^4.9.4",
+    "vitest": "^0.28.4"
+  },
+  "type": "module"
 }
