@@ -1,40 +1,46 @@
 {
-  "name": "team-mates-ui",
-  "version": "0.0.1",
-  "author": "thomascal",
-  "license": "MIT",
+  "name": "team-mates",
+  "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --open",
-    "mock": "ng serve --open --configuration mock",
+    "start": "ng serve",
     "build": "ng build",
-    "watch": "ng build --watch --configuration development"
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.1.1",
-    "@angular/cdk": "^13.1.1",
-    "@angular/common": "~13.1.1",
-    "@angular/compiler": "~13.1.1",
-    "@angular/core": "~13.1.1",
-    "@angular/forms": "~13.1.1",
-    "@angular/platform-browser": "~13.1.1",
-    "@angular/platform-browser-dynamic": "~13.1.1",
-    "@holochain/client": "^0.3.1",
-    "@ng-select/ng-select": "^8.1.1",
-    "@ngrx/component-store": "^13.0.2",
-    "ngx-image-cropper": "^6.0.2",
-    "resize-observer": "^1.0.4",
-    "rxjs": "~7.4.0",
-    "tslib": "^2.3.1",
-    "zone.js": "~0.11.4"
+    "@angular/animations": "^15.2.8",
+    "@angular/common": "^15.2.8",
+    "@angular/compiler": "^15.2.8",
+    "@angular/core": "^15.2.8",
+    "@angular/forms": "^15.2.8",
+    "@angular/platform-browser": "^15.2.8",
+    "@angular/platform-browser-dynamic": "^15.2.8",
+    "@angular/router": "^15.2.8",
+    "@holochain-open-dev/profiles": "0.14.1",
+    "@holochain/client": "0.12.5",
+    "@msgpack/msgpack": "^3.0.0-beta2",
+    "@ng-select/ng-select": "^10.0.4",
+    "@ngrx/component-store": "^15.4.0",
+    "ngx-image-cropper": "^6.3.3",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.1.2",
-    "@angular/cli": "~13.1.2",
-    "@angular/compiler-cli": "~13.1.1",
-    "@types/node": "^17.0.0",
-    "tailwindcss": "^3.0.7",
-    "typescript": "~4.5.4"
+    "@angular-devkit/build-angular": "^15.2.7",
+    "@angular/cli": "~15.2.7",
+    "@angular/compiler-cli": "^15.2.8",
+    "autoprefixer": "^10.4.14",
+    "postcss": "^8.4.23",
+    "tailwindcss": "^3.3.2",
+    "typescript": "~4.9.5"
+  },
+  "browser": {
+    "crypto": false
+  },
+  "engines": {
+    "npm": ">=8.0.0 <10.0.0",
+    "node": ">=16.0.0 <19.0.0"
   }
 }
