@@ -1,11 +1,14 @@
 import { Component} from "@angular/core";
 import { Observable } from "rxjs";
 import { KeyValue, Profile } from "../../../models/profile"
-import { map } from 'rxjs/operators';
-import { ProfileStore } from '../../../store/profile.store';
+import { map, filter } from 'rxjs/operators';
+import { ProfileStore } from '../../../stores/profile.store';
+import { CommonModule } from "@angular/common";
 
 @Component({
   selector: "all-profiles",
+  standalone: true,
+  imports: [CommonModule],
   templateUrl: "./all-profiles.component.html",
 })
 export class AllProfilesComponent {
@@ -16,9 +19,18 @@ export class AllProfilesComponent {
   onenick: string = ""
 
   constructor(private _profileStore: ProfileStore){
-    this.profiles$  = this._profileStore.selectAllProfiles()
-      .pipe(map(plist=>plist.filter(pro=>pro.agentPubKey !== this._profileStore.mypubkey)
-      .map(plist=> {return plist.profile!})));
+    this.profiles$  = this._profileStore.selectOtherProfiles()//.selectAllAgentProfiles()
+    //.pipe(filter((profiles: Profile[])  => { return .agentPubKey64 !== this._profileStore.mypubkey64)
+      //=> agentprofiles.map(ap => {
+    //.pipe(map(plist => { return plist.filter( fap => fap.agentPubKey64 !== this._profileStore.mypubkey64)
+      //      .map(ap=>ap.profile)})
+        //console.log(JSON.stringify(ap.agentPubKey))
+        //console.log(JSON.stringify(this._profileStore.mypubkey))
+     
+     // } 
+        //return plist.filter(pro=>JSON.stringify(pro.agentPubKey) !== JSON.stringify(this._profileStore.mypubkey))
+      //.map(p=> {return p.profile})}
+      //);
   }
 
   showprofile(profile:Profile){
