@@ -1,13 +1,17 @@
 import { ChangeDetectionStrategy, Component, OnDestroy } from '@angular/core';
 import { InvitationEntryInfo } from 'src/app/models/invitation';
-import { InvitationStore } from '../../../store/invitation.store';
-import { ProfileStore } from 'src/app/store/profile.store';
+import { InvitationStore } from 'src/app/stores/invitation.store';
+import { ProfileStore } from 'src/app/stores/profile.store';
 import { map } from 'rxjs/operators';
 import { Dictionary } from 'src/app/helpers/utils';
 import { Observable, Subscription } from 'rxjs';
+import { ActionHash, AgentPubKey, EntryHash } from '@holochain/client';
+import { CommonModule } from '@angular/common';
 
 @Component({
   selector: 'list-invitations',
+  standalone: true,
+  imports: [CommonModule],
   templateUrl: './list-invitations.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
@@ -32,27 +36,27 @@ export class InvitationListComponent implements OnDestroy{
     this.profileDictionarySubscription$.unsubscribe()
   }
 
-  accept(entry_hash:string){
-    this._invitationStore.acceptInvitation(entry_hash)
+  accept(action_hash:ActionHash){
+    this._invitationStore.acceptInvitation(action_hash)
   }
 
-  reject(entry_hash:string){
-    this._invitationStore.rejectInvitation(entry_hash)
+  reject(action_hash:ActionHash){
+    this._invitationStore.rejectInvitation(action_hash)
   }
 
-  cancel(entry_hash:string){
+  cancel(entry_hash:EntryHash){
     this._invitationStore.clearInvitation(entry_hash)
   }
 
-  hashToAgent(hash:string):string{
+  hashToAgent(hash:AgentPubKey):string{
     //console.log("profile_list:",this.profileDictionary)
-    return this.profileDictionary[hash]
+    return this.profileDictionary[hash.join()]
   }
 
-  hashListToAgentList(hashlist:string[]):string[] {
+  hashListToAgentList(hashlist:AgentPubKey[]):string[] {
     const res:string[] = []
     for(const hash of hashlist){
-      res.push(" "+this.profileDictionary[hash])
+      res.push(" "+this.profileDictionary[hash.join()])
     }
     return res
   }
