@@ -1,5 +1,5 @@
 import { Dictionary } from "../helpers/utils";
-import { AgentPubKeyB64, HeaderHashB64, EntryHashB64 } from '@holochain-open-dev/core-types';
+import { AgentPubKey, ActionHash, EntryHash, AgentPubKeyB64 } from '@holochain/client'
 
 export interface Profile {
   fields: Dictionary<string>;
@@ -7,13 +7,15 @@ export interface Profile {
 }
 
 export interface AgentProfile {
-  agentPubKey: AgentPubKeyB64;
-  profile: Profile | undefined;
+  agentPubKey: AgentPubKey
+  agentPubKey64: AgentPubKeyB64
+  keyNick: String
+  profile: Profile // | undefined;
 }
 
 export interface KeyNick {
-  agent_pub_key:string;
-  nickname: string
+  agent_pub_key:AgentPubKeyB64;
+  nickname: String
 }
 
 export interface KeyValue {
@@ -26,24 +28,26 @@ const default_avatar = "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAMgAAADICA
 
 export const mockMyProfile:Profile = {nickname: "thomas",fields:{["avatar"]:""}}
 
-export const mockMyAgentProfile:AgentProfile = {agentPubKey: "12C0kP3Cu8QRxERdKJZIqlI3y_gQuJke5qFp7Ae52L49N-vs", profile: mockMyProfile};
+export const mockMyAgentProfile:AgentProfile = {agentPubKey: new Uint8Array(10), agentPubKey64:"64stringhere", keyNick:mockMyProfile.nickname+"1ddw", profile: mockMyProfile};
 
 
 export const mock1Profile:Profile = {nickname: "friend1",fields:{["avatar"]:default_avatar,["age"]:"21",["nationality"]:"afgan"}}
 
-export const mock1AgentProfile:AgentProfile = {agentPubKey: "13C0kP3Cu8QRxERdKJZIqlI3y_gQuJke5qFp7Ae52L49N-ds", profile: mock1Profile};
+export const mock1AgentProfile:AgentProfile = {agentPubKey: new Uint8Array(11), agentPubKey64:"64stringhere",keyNick:mockMyProfile.nickname+"fwef", profile: mock1Profile};
 
 export const mock2Profile:Profile = {nickname: "friend2",fields:{["avatar"]:default_avatar}}
 
-export const mock2AgentProfile:AgentProfile = {agentPubKey: "13C0kP3Cu8QRxERdKJZIqlI3y_gQuJke5qFp7Ae52L49N-es", profile: mock2Profile};
+export const mock2AgentProfile:AgentProfile = {agentPubKey: new Uint8Array(11), agentPubKey64:"64stringhere",keyNick:mockMyProfile.nickname+"3565", profile: mock2Profile};
 
 export const mock3Profile:Profile = {nickname: "friend3",fields:{["avatar"]:default_avatar}}
 
-export const mock3AgentProfile:AgentProfile = {agentPubKey: "13C0kP3Cu8QRxERdKJZIqlI3y_gQuJke5qFp7Ae52L49N-fs", profile: mock3Profile};
+export const mock3AgentProfile:AgentProfile = {agentPubKey: new Uint8Array(11), agentPubKey64:"64stringhere",keyNick:mockMyProfile.nickname+"wfsg", profile: mock3Profile};
 
 export const mock4Profile:Profile = {nickname: "friend4",fields:{["avatar"]:default_avatar}}
 
-export const mock4AgentProfile:AgentProfile = {agentPubKey: "13C0kP3Cu8QRxERdKJZIqlI3y_gQuJke5qFp7Ae52L49N-gs", profile: mock4Profile};
+export const mock4AgentProfile:AgentProfile = {agentPubKey: new Uint8Array(11), agentPubKey64:"64stringhere",keyNick:mockMyProfile.nickname+"4GGD", profile: mock4Profile};
 
 export const mockAgentProfiles:AgentProfile[] = [mock1AgentProfile,mock2AgentProfile,mock3AgentProfile,mock4AgentProfile];
+
+export const mockAgents:AgentPubKey[] = [mock1AgentProfile.agentPubKey,mock2AgentProfile.agentPubKey,mock3AgentProfile.agentPubKey,mock4AgentProfile.agentPubKey]
 
