@@ -2,12 +2,13 @@
 // `ng build` replaces `environment.ts` with `environment.prod.ts`.
 // The list of file replacements can be found in `angular.json`.
 
+
 export const environment = {
   production: false,
   mock: false,
   HOST_URL: 'ws://127.0.0.1:8888',
-  APP_ID:'team-mates',
-  cell1:'profile_invitation'
+  APP_ID:'team-mates-app',
+  cells: ['team-mates']
 };
 
 /*
