@@ -1,6 +1,6 @@
+/** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: ['./src/**/*.{html,ts}'],
-  //darkMode: false, // or 'media' or 'class'
+  content: ["./src/**/*.{html,ts}"],
   theme: {
     extend: {
       backgroundImage: {
@@ -12,8 +12,6 @@ module.exports = {
       }
     },
   },
-  variants: {
-    extend: {},
-  },
   plugins: [],
 }
+
